misc/HtmlTestSuite/results/test-1019.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Meta color-scheme in body should apply</title>
<link href="https://html.spec.whatwg.org/multipage/semantics.html#meta-color-scheme" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="support/compute-root-color-scheme.js"/>
</head><body>
  <meta content="dark" name="color-scheme"/>
<script>
  assert_root_color_scheme("dark", "Meta color-scheme in body should apply.");
</script>
</body>
</html>
